#194 [REFACTOR] : onboarding 패키지명 소문자로 수정
