+++ app/src/main/java/com/example/charo_android/presentation/ui/onboarding/OnBoardingOneFragment.kt
@@ -1,4 +1,4 @@
-package com.example.charo_android.presentation.ui.onBoarding
+package com.example.charo_android.presentation.ui.onboarding
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,26 +6,26 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.charo_android.databinding.FragmentOnBoarding2Binding
-import com.example.charo_android.presentation.ui.signin.SignInActivity
+import com.example.charo_android.databinding.FragmentOnBoarding1Binding
 import com.example.charo_android.presentation.ui.signin.SocialSignInActivity
 
-class OnBoardingTwoFragment : Fragment() {
-    private var _binding: FragmentOnBoarding2Binding? = null
-    private val binding get() = _binding!!
 
+class OnBoardingOneFragment : Fragment() {
+    private var _binding: FragmentOnBoarding1Binding? = null
+    private val binding get() = _binding!!
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentOnBoarding2Binding.inflate(inflater, container, false)
-
+        _binding = FragmentOnBoarding1Binding.inflate(inflater, container, false)
 
         binding.btnSkips.setOnClickListener {
             val intent = Intent(requireActivity(), SocialSignInActivity::class.java)
             startActivity(intent)
         }
-        return binding.root
 
+        return binding.root
     }
+
+
 }
