+++ app/src/main/java/com/example/charo_android/presentation/ui/onboarding/OnBoardingThreeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.charo_android.presentation.ui.onBoarding
+package com.example.charo_android.presentation.ui.onboarding
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,7 +7,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.charo_android.databinding.FragmentOnBoarding3Binding
-import com.example.charo_android.presentation.ui.signin.SignInActivity
 import com.example.charo_android.presentation.ui.signin.SocialSignInActivity
