+++ app/src/main/res/layout/activity_on_boarding.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.onBoarding.OnBoardingActivity">
+    tools:context=".presentation.ui.onboarding.OnBoardingActivity">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/vp_onboarding"
