+++ app/src/main/res/layout/fragment_on_boarding_1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.onBoarding.OnBoardingOneFragment">
+    tools:context=".presentation.ui.onboarding.OnBoardingOneFragment">
 
     <Button
         android:id="@+id/btn_skips"
