+++ app/src/main/res/layout/fragment_on_boarding_2.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.ui.onBoarding.OnBoardingTwoFragment">
+    tools:context=".presentation.ui.onboarding.OnBoardingTwoFragment">
 
     <Button
         android:id="@+id/btn_skips"
