+++ app/src/main/res/layout/fragment_on_boarding_3.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.onBoarding.OnBoardingThreeFragment">
+    tools:context=".presentation.ui.onboarding.OnBoardingThreeFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
